requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
openai==1.54.0
pydantic==2.10.3
python-multipart==0.0.9
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.8.3
groq==0.9.0
